% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_and_commit.R
\name{add_and_commit}
\alias{add_and_commit}
\title{This function adds files to the git tracking list to solve the untracked files problem.}
\usage{
add_and_commit(
  files,
  repository = "Frost2021Package",
  message = "Working on it",
  commit = TRUE
)
}
\arguments{
\item{files, }{the files}

\item{repository, }{the repository}

\item{message, }{commit message}
}
\description{
This function adds files to the git tracking list to solve the untracked files problem.
}
