% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_package.R
\name{add_to_package}
\alias{add_to_package}
\title{This function adds code to the project R package.}
\usage{
add_to_package(file)
}
\arguments{
\item{file, }{the file to be added}
}
\description{
This function adds code to the project R package.
}
