% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframe.R
\name{build_dataframe}
\alias{build_dataframe}
\title{Builds the Dataframe}
\usage{
build_dataframe(url, css_selectors)
}
\arguments{
\item{url, }{url to be scraped}

\item{css_selectors, }{vector of css_selectors to scrape}
}
\value{
dataframe with scraped data organized into columns by their CSS selector tag. May require postprocessing.
}
\description{
Builds the Dataframe
}
