% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mdf.R
\name{build_mdf}
\alias{build_mdf}
\title{This function builds the metadata dataframe from a given list of data frames.}
\usage{
build_mdf(arr)
}
\arguments{
\item{arr, }{list of data frames}
}
\value{
complete metadata dataframe
}
\description{
This function builds the metadata dataframe from a given list of data frames.
}
