% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNull.R
\name{checkNull}
\alias{checkNull}
\title{This function checks if a variable is equivalent to character(0) and handles the data accordingly.}
\usage{
checkNull(item)
}
\arguments{
\item{item}{the item to be checked}
}
\value{
NA if the item is equivalent to character(0), the item otherwise
}
\description{
This function checks if a variable is equivalent to character(0) and handles the data accordingly.
}
