% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{This function cleans the data and removes unnecessary columns}
\usage{
clean(df)
}
\arguments{
\item{df, }{data frame}
}
\value{
cleaned dataset
}
\description{
This function cleans the data and removes unnecessary columns
}
