% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_vec.R
\name{col_to_vec}
\alias{col_to_vec}
\title{This converts a data frame column to a vector, similar to the "$" operator in base R.}
\usage{
col_to_vec(col)
}
\arguments{
\item{col, }{the data frame column}
}
\value{
vector with elements of the column
}
\description{
This converts a data frame column to a vector, similar to the "$" operator in base R.
}
