% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{This function converts the file size to MB.}
\usage{
convert(x)
}
\arguments{
\item{x}{the file size in string and unit format}
}
\value{
the file size in MB
}
\description{
This function converts the file size to MB.
}
