% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{This function copies and moves files}
\usage{
copy(
  filename,
 
    path = "/Users/sucheen/Documents/Cal Poly SLO/Academics/Summer 2021/Frost2021Package/R/"
)
}
\arguments{
\item{filename, }{the file to be copied}

\item{path, }{the destination path}
}
\description{
This function copies and moves files
}
