% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_rowid.R
\name{drop_rowid}
\alias{drop_rowid}
\title{This function drops the rowid columns}
\usage{
drop_rowid(df)
}
\arguments{
\item{df, }{data frame}
}
\value{
df without rowid columns
}
\description{
This function drops the rowid columns
}
