% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Optimizes vector indexing for the magrittr package.}
\usage{
get(vec, index)
}
\arguments{
\item{vec}{vector of elements}

\item{index}{the vector index}
}
\value{
element at the index
}
\description{
Optimizes vector indexing for the magrittr package.
}
