% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{This function does the git commit and push process.}
\usage{
git(repository = "Frost2021Package", message = "Working on it")
}
\arguments{
\item{repository, }{the R package}

\item{message, }{commit message}
}
\description{
This function does the git commit and push process.
}
