% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{links}
\alias{links}
\title{This function scrapes data links according to a given CSS selector.}
\usage{
links(url, css)
}
\arguments{
\item{url}{the website url}
}
\value{
vector of links
}
\description{
This function scrapes data links according to a given CSS selector.
}
