% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_vec.R
\name{list_to_vec}
\alias{list_to_vec}
\title{This function converts a list to a vector.}
\usage{
list_to_vec(l)
}
\arguments{
\item{l, }{the list}
}
\value{
vector with the elements of the list
}
\description{
This function converts a list to a vector.
}
