% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_drop.R
\name{merge_and_drop}
\alias{merge_and_drop}
\title{This function perfoms the merge-and-drop technique for condensing the metadata dataframe as discussed at a meeting. Shifts from col2 to col1.}
\usage{
merge_and_drop(df, col1, col2)
}
\arguments{
\item{df, }{data frame}

\item{col1, }{the first column}

\item{col2, }{the second column}
}
\value{
merged dataframe
}
\description{
This function perfoms the merge-and-drop technique for condensing the metadata dataframe as discussed at a meeting. Shifts from col2 to col1.
}
