% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cdph.R
\name{scrape_cdph}
\alias{scrape_cdph}
\title{This function scrapes metadata attributes from the CDC metadata repository.}
\usage{
scrape_cdph(url)
}
\arguments{
\item{url}{the website url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the CDC metadata repository.
}
