% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_datagov.R
\name{scrape_datagov}
\alias{scrape_datagov}
\title{This function scrapes metadata attributes from the data.gov repository.}
\usage{
scrape_datagov(url)
}
\arguments{
\item{url}{the website url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the data.gov repository.
}
