% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_datahub.R
\name{scrape_datahub}
\alias{scrape_datahub}
\title{Extracts metadata attributes from Datahub.io}
\usage{
scrape_datahub(url)
}
\arguments{
\item{url, }{the website url (must be linked to datahub.io)}
}
\value{
dataframe of metadata attributes
}
\description{
Extracts metadata attributes from Datahub.io
}
