% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_datashare.R
\name{scrape_datashare}
\alias{scrape_datashare}
\title{This function scrapes metadata attributes from the edinburgh datashare repository.}
\usage{
scrape_datashare(url)
}
\arguments{
\item{url, }{the url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the edinburgh datashare repository.
}
