% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_dryad.R
\name{scrape_dryad}
\alias{scrape_dryad}
\title{This function scrapes metadata (date uploaded, viewing metrics, etc.) from the Data Dryad repository.}
\usage{
scrape_dryad(url)
}
\arguments{
\item{url, }{the website url}
}
\value{
dataframe with scraped metadata contents from Data Dryad
}
\description{
This function scrapes metadata (date uploaded, viewing metrics, etc.) from the Data Dryad repository.
}
