% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_figshare.R
\name{scrape_figshare}
\alias{scrape_figshare}
\title{This function scrapes metadata attributes from the Figshare metadata repository.}
\usage{
scrape_figshare(url)
}
\arguments{
\item{url}{the website url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the Figshare metadata repository.
}
