% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_harvard.R
\name{scrape_harvard}
\alias{scrape_harvard}
\title{This function scrapes metadata attributes from the Harvard Dataverse metadata repository.}
\usage{
scrape_harvard(url)
}
\arguments{
\item{url}{the website url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the Harvard Dataverse metadata repository.
}
