% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rvest.R
\name{scrape_rvest}
\alias{scrape_rvest}
\title{Returns scraped data based on given URL and CSS selector}
\usage{
scrape_rvest(url, type)
}
\arguments{
\item{url}{the website url}

\item{type}{the html formatting type to extract the correct items}
}
\description{
Returns scraped data based on given URL and CSS selector
}
\author{
Sucheen Sundaram
Date 6/21/2021
}
