% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_zenodo.R
\name{scrape_zenodo}
\alias{scrape_zenodo}
\title{This function scrapes metadata attributes from the Zenodo data repository.}
\usage{
scrape_zenodo(url)
}
\arguments{
\item{url}{the website url}
}
\value{
dataframe with metadata attributes
}
\description{
This function scrapes metadata attributes from the Zenodo data repository.
}
