% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_score.R
\name{similarity_score}
\alias{similarity_score}
\title{This function calculates a similarity score between two strings.}
\usage{
similarity_score(a, b)
}
\arguments{
\item{a, }{string A}

\item{b, }{string B}
}
\value{
similarity score
}
\description{
This function calculates a similarity score between two strings.
}
