% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{This function modifies the dataframe based on a set of target columns.
It will replace a column with NA if it is not present in the dataframe and only
return those columns.}
\usage{
target(df, target_cols)
}
\arguments{
\item{df, }{data frame}

\item{target, }{target columns}
}
\value{
data frame modified around the target columns as specified
}
\description{
This function modifies the dataframe based on a set of target columns.
It will replace a column with NA if it is not present in the dataframe and only
return those columns.
}
