% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_na.R
\name{unique_na}
\alias{unique_na}
\title{This returns distinct values of the elements in a vector, including NA values.}
\usage{
unique_na(df, column)
}
\arguments{
\item{df}{data frame}

\item{column}{the column}
}
\value{
the distinct values of the column
}
\description{
This returns distinct values of the elements in a vector, including NA values.
}
