% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_load.R
\name{update_load}
\alias{update_load}
\title{This function updates the namespace for loading into other rproj files.}
\usage{
update_load(package = "Frost2021Package")
}
\description{
This function updates the namespace for loading into other rproj files.
}
\author{
Sucheen Sundaram
}
