% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_eq.R
\name{vec_eq}
\alias{vec_eq}
\title{This function checks if two vectors are equal, that is, if they are the same length and their elements
are in the same order.}
\usage{
vec_eq(a, b)
}
\arguments{
\item{a, }{vector a}

\item{b, }{vector b}
}
\value{
whether or not a and b are equal according to the above definition
}
\description{
This function checks if two vectors are equal, that is, if they are the same length and their elements
are in the same order.
}
